app/requirements.txt
# FastAPI and Web Framework
fastapi==0.120.0
uvicorn==0.38.0
python-multipart==0.0.20
fastapi[standard]


# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# Data Validation
pydantic==2.12.3
pydantic-settings==2.11.0

# Authentication and Security
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.1.1
cryptography==46.0.3
